Crate bholdus_runtime

Re-exports

pub use constants::currency::*;

pub use constants::time::*;

pub use pallet_template;

Modules

api

constants
Import some useful constants A set of constant values used in substrate runtime.

opaque
Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

weights

Macros

construct_runtime
Construct a runtime, with the given name and the given pallets.

parameter_types
Create new implementations of the Get trait.

Structs

AnnouncementDepositBase

AnnouncementDepositFactor

ApprovalDeposit

AssetDeposit

BasicDeposit

BlockExecutionWeight
Importing a block with 0 txs takes ~5 ms

BlockHashCount

BondingDuration

BountyCuratorDeposit

BountyDepositBase

BountyDepositPayoutDelay

BountyUpdatePeriod

BountyValueMinimum

BridgeProposalLifetime

Burn

ChainId

ConfigDepositBase

CouncilMaxMembers

CouncilMaxProposals

CouncilMotionDuration

DataDepositPerByte

DeletionQueueDepth

DeletionWeightLimit

DepositBase

DepositFactor

DepositPerContract

DepositPerStorageByte

DepositPerStorageItem

DexPalletId

DisabledValidatorsThreshold

EpochDuration

ExchangeFee

ExistentialDeposit

ExistentialDeposits

ExpectedBlockTime

ExtrinsicBaseWeight
Executing 10,000 System remarks (no-op) txs takes ~1.26 seconds -> ~125 µs per tx

Fallback

FieldDeposit

FriendDepositFactor

GenesisConfig

GetNativeCurrencyId

IdentityFee
Implementor of WeightToFeePolynomial that maps one unit of weight to one unit of fee.

ImOnlineUnsignedPriority

IndexDeposit

MaxAdditionalFields

MaxApprovals

MaxAttributesBytes

MaxClassMetadata

MaxDepth

MaxFriends

MaxLocks

MaxNominatorRewardedPerValidator

MaxPending

MaxProxies

MaxRegistrars

MaxScheduledPerBlock

MaxSignatories

MaxSubAccounts

MaxTokenMetadata

MaxValueSize

MaximumReasonLength

MaximumSchedulerWeight

MetadataDepositBase

MetadataDepositPerByte

MinerMaxIterations

MinerMaxLength

MinerMaxWeight

MinimumPeriod

MultiPhaseUnsignedPriority

NftPalletId

NposCompactSolution16

OffchainRepeat

OffencesWeightSoftLimit

Offset

Origin

PalletId
A pallet identifier. These are per pallet and should be stored in a registry somewhere.

PalletInfo
Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Period

ProposalBond

ProposalBondMinimum

ProxyDepositBase

ProxyDepositFactor

RecoveryDeposit

RentFraction

ReportLongevity

RewardCurve

RocksDbWeight
By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

SS58Prefix

Schedule

SessionsPerEra

SignedClaimHandicap

SignedDepositBase

SignedDepositByte

SignedMaxSubmissions

SignedPhase

SignedRewardBase

SlashDeferDuration

SolutionImprovementThreshold

SpendPeriod

StakingUnsignedPriority
We prioritize im-online heartbeats over election solution submission.

StringLimit

SubAccountDeposit

SurchargeReward

TipCountdown

TipFindersFee

TipReportDepositBase

TombstoneDeposit

TradingPathLimit

TransactionByteFee

TreasuryPalletId

U128CurrencyToVote
An implementation of CurrencyToVote tailored for chain’s that have a balance type of u128.

UncleGenerations

UnsignedPhase

Version

Enums

BalancesCall
Contains one variant per dispatchable that can be called by an extrinsic.

Call

DispatchClass
A generalized group of dispatch types.

Event

OriginCaller

ProxyType
The type used to represent the kinds of proxying allowed.

StakerStatus
Indicates the initial status of the staker.

TimestampCall
Contains one variant per dispatchable that can be called by an extrinsic.

Constants

MAX_NOMINATIONS

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

WEIGHT_PER_SECOND

Traits

BuildStorage
Complex storage builder stuff.

InstanceFilter
Simple trait for providing a filter over a reference to some type, given an instance of itself.

KeyOwnerProofSystem
Something which can compute and check proofs of a historical key owner and return full identification data of that key owner.

MaxEncodedLen
Items implementing MaxEncodedLen have a statically known maximum encoded size.

Randomness
A trait that is able to provide randomness.

StorageValue
A trait for working with macro-generated storage values under the substrate storage API.

Functions

native_version
The version information used to identify this runtime when compiled natively.

Type Definitions

Address
The address format for describing accounts.

AllModules	Deprecated
All modules included in the runtime as a nested tuple of types. Excludes the System pallet.

AllModulesWithSystem	Deprecated
All modules included in the runtime as a nested tuple of types.

AllPallets
All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithSystem
All pallets included in the runtime as a nested tuple of types.

Aura

AuraConfig

AuthorityDiscovery

AuthorityDiscoveryConfig

Authorship

Balances

BalancesConfig

BholdusLibNFT

Block
Block type as expected by this runtime.

Bounties

ChainBridge

ChainBridgeTransfer

ChainBridgeTransferConfig

Contracts

Council

CouncilConfig

Currencies

Dex

DexConfig

ElectionProviderMultiPhase

Executive
Executive: handles dispatch to the various modules.

Grandpa

GrandpaConfig

Header
Block header type as expected by this runtime.

Historical

Identity

ImOnline

ImOnlineConfig

Indices

IndicesConfig

Mmr

Multisig

NFT

Offences

Proxy

RandomnessCollectiveFlip

Recovery

Scheduler

Session

SessionConfig

SignedExtra
The SignedExtension to the basic transaction logic.

SignedPayload
The payload being signed in transactions.

Staking

StakingConfig

Sudo

SudoConfig

System

SystemConfig

TemplateModule

Timestamp

Tokens

TokensConfig

TransactionPayment

Treasury

TreasuryConfig

UncheckedExtrinsic
Unchecked extrinsic type as expected by this runtime.

Utility

Weight
Numeric range of a transaction weight.

Derive Macros

MaxEncodedLen
Derive MaxEncodedLen.

RuntimeDebug