Trait bholdus_runtime::InstanceFilter

pub trait InstanceFilter<T>: Send + Sync {
    pub fn filter(&self, &T) -> bool;

    pub fn is_superset(&self, _o: &Self) -> bool { ... }
}

Simple trait for providing a filter over a reference to some type, given an instance of itself.

Required methods

pub fn filter(&self, &T) -> bool

Determine if a given value should be allowed through the filter (returns true) or not.

Loading content...
Provided methods

pub fn is_superset(&self, _o: &Self) -> bool

Determines whether self matches at least everything that _o does.

Loading content...